memory: add fixed height for image to fix go to anchor
Because figure automatically use lazy loading, the browser can't know
the actual height of the images and this was breaking going straight to
any section that was after an image. Also use object-fit: contain on the
img so that mobiles keep the aspect ratio at the cost of loosing some
unnecessary vertical space.